[CI] Organize speculative decoding E2E tests by coverage (#50330)
