+++ .buildkite/intel_jobs/engine_intel.yaml
@@ -95,9 +95,11 @@ steps:
     - vllm/triton_utils/
     - vllm/utils/
     - vllm/v1/
-    - tests/v1/e2e/spec_decode
+    - tests/v1/e2e/spec_decode/
   commands:
   - >-
       bash .buildkite/scripts/hardware_ci/run-intel-test.sh
       'cd tests &&
-      pytest -v -s v1/e2e/spec_decode/test_spec_decode.py -k "tensor_parallelism"'
+      pytest -v -s
+      v1/e2e/spec_decode/draft_model/test_draft_model.py::test_draft_model_tensor_parallelism
+      v1/e2e/spec_decode/draft_model/test_draft_model.py::test_draft_model_engine_args_tensor_parallelism'
