+++ .buildkite/intel_jobs/model_runner_v2_intel.yaml
@@ -107,7 +107,6 @@ steps:
   - vllm/v1/worker/gpu_worker.py
   - tests/v1/spec_decode/test_max_len.py
   - tests/v1/spec_decode/test_rejection_sampler_utils.py
-  - tests/v1/e2e/spec_decode/test_spec_decode.py
   commands:
     - >-
       bash .buildkite/scripts/hardware_ci/run-intel-test.sh
