+++ .buildkite/test_areas/spec_decode.yaml
@@ -2,16 +2,43 @@ group: Spec Decode
 depends_on:
   - image-build
 steps:
+- label: V1 Spec Decode
+  device: h200_35gb
+  key: v1-spec-decode
+  timeout_in_minutes: 40
+  source_file_dependencies:
+    - vllm/config/
+    - vllm/distributed/
+    - vllm/inputs/
+    - vllm/model_executor/
+    - vllm/platforms/
+    - vllm/sampling_params.py
+    - vllm/transformers_utils/
+    - vllm/utils/
+    - vllm/v1/
+    - tests/v1/spec_decode
+  commands:
+    - export VLLM_WORKER_MULTIPROC_METHOD=spawn
+    # TODO: create another `optional` test group for slow tests
+    - pytest -v -s -m 'not slow_test' v1/spec_decode
+  mirror:
+    amd:
+      dind: false
+      device: mi300_1
+      timeout_in_minutes: 50
+      depends_on:
+      - image-build-amd
+
 - label: Spec Decode Eagle
   key: spec-decode-eagle
   timeout_in_minutes: 25
-  device: h200_18gb
+  device: h200_35gb
   source_file_dependencies:
     - vllm/v1/spec_decode/
     - vllm/v1/worker/gpu/spec_decode/
     - tests/v1/e2e/spec_decode/
   commands:
-    - pytest -v -s v1/e2e/spec_decode -k "eagle_correctness"
+    - pytest -v -s v1/e2e/spec_decode/eagle/
   mirror:
     amd:
       dind: false
@@ -28,7 +55,7 @@ steps:
       - tests/v1/e2e/spec_decode/
       - vllm/platforms/rocm.py
 
-- label: Spec Decode Eagle Nightly B200
+- label: Spec Decode Eagle Nightly (B200)
   key: spec-decode-eagle-nightly-b200
   timeout_in_minutes: 25
   device: b200-k8s
@@ -38,12 +65,12 @@ steps:
     - vllm/v1/worker/gpu/spec_decode/
     - tests/v1/e2e/spec_decode/
   commands:
-    - pytest -v -s v1/e2e/spec_decode -k "eagle_correctness"
+    - pytest -v -s v1/e2e/spec_decode/eagle/
 
 - label: Spec Decode Speculators + MTP
   key: spec-decode-speculators-mtp
-  timeout_in_minutes: 20
-  device: h200_18gb
+  timeout_in_minutes: 50
+  device: h200_35gb
   source_file_dependencies:
     - vllm/v1/spec_decode/
     - vllm/v1/worker/gpu/spec_decode/
@@ -51,7 +78,8 @@ steps:
     - vllm/transformers_utils/configs/speculators/
     - tests/v1/e2e/spec_decode/
   commands:
-    - pytest -v -s v1/e2e/spec_decode -k "speculators or mtp_correctness"
+    - pytest -v -s v1/e2e/spec_decode/speculators/
+    - pytest -v -s v1/e2e/spec_decode/mtp/
   mirror:
     amd:
       dind: false
@@ -69,7 +97,7 @@ steps:
       - tests/v1/e2e/spec_decode/
       - vllm/platforms/rocm.py
 
-- label: Spec Decode Speculators + MTP Nightly B200
+- label: Spec Decode Speculators + MTP Nightly (B200)
   key: spec-decode-speculators-mtp-nightly-b200
   timeout_in_minutes: 30
   device: b200-k8s
@@ -80,19 +108,20 @@ steps:
     - vllm/transformers_utils/configs/speculators/
     - tests/v1/e2e/spec_decode/
   commands:
-    - pytest -v -s v1/e2e/spec_decode -k "speculators or mtp_correctness"
-  
+    - pytest -v -s v1/e2e/spec_decode/speculators/
+    - pytest -v -s v1/e2e/spec_decode/mtp/
+
 - label: Spec Decode Ngram + Suffix
   key: spec-decode-ngram-suffix
   timeout_in_minutes: 20
-  device: h200_18gb
+  device: h200_35gb
   source_file_dependencies:
     - vllm/v1/spec_decode/
     - vllm/v1/worker/gpu/spec_decode/
     - tests/v1/e2e/spec_decode/
     - tests/spec_decode/
   commands:
-    - pytest -v -s v1/e2e/spec_decode -k "ngram or suffix"
+    - pytest -v -s v1/e2e/spec_decode/ngram_suffix/
     - python3 spec_decode/test_custom_proposer.py
   mirror:
     amd:
@@ -119,7 +148,7 @@ steps:
     - vllm/v1/worker/gpu/spec_decode/
     - tests/v1/e2e/spec_decode/
   commands:
-    - pytest -v -s v1/e2e/spec_decode -k "draft_model or no_sync or batch_inference"
+    - pytest -v -s v1/e2e/spec_decode/draft_model/
   mirror:
     amd:
       dind: false
@@ -136,7 +165,7 @@ steps:
       - tests/v1/e2e/spec_decode/
       - vllm/platforms/rocm.py
 
-- label: Spec Decode Draft Model Nightly B200
+- label: Spec Decode Draft Model Nightly (B200)
   key: spec-decode-draft-model-nightly-b200
   timeout_in_minutes: 40
   device: b200-k8s
@@ -146,9 +175,9 @@ steps:
     - vllm/v1/worker/gpu/spec_decode/
     - tests/v1/e2e/spec_decode/
   commands:
-    - pytest -v -s v1/e2e/spec_decode -k "draft_model or no_sync or batch_inference"
+    - pytest -v -s v1/e2e/spec_decode/draft_model/
 
-- label: Speculators Correctness
+- label: Speculators Correctness Nightly (H100)
   key: speculators-correctness
   timeout_in_minutes: 30
   device: h100
@@ -156,24 +185,12 @@ steps:
   num_devices: 1
   source_file_dependencies:
     - vllm/v1/spec_decode/
+    - vllm/v1/worker/gpu/spec_decode/dflash/
     - vllm/model_executor/models/qwen3_dflash.py
     - tests/v1/spec_decode/test_speculators_correctness.py
   commands:
     - export VLLM_ALLOW_INSECURE_SERIALIZATION=1
     - pytest -v -s v1/spec_decode/test_speculators_correctness.py -m slow_test
-
-- label: Spec Decode MTP hybrid (B200)
-  timeout_in_minutes: 20
-  device: b200-k8s
-  optional: true
-  source_file_dependencies:
-    - vllm/v1/spec_decode/
-    - vllm/v1/worker/gpu/spec_decode/
-    - vllm/model_executor/models/qwen3_5.py
-    - vllm/model_executor/models/qwen3_5_mtp.py
-    - tests/v1/e2e/spec_decode/
-  commands:
-    - pytest -v -s v1/e2e/spec_decode -k "qwen3_5-hybrid"
 
 - label: Spec Decode DeepSeek MTP Parallel Load (B200)
   key: spec-decode-deepseek-mtp-parallel-load-b200
@@ -191,14 +208,43 @@ steps:
   commands:
     - pytest -v -s v1/e2e/spec_decode/test_mtp_parallel_load.py
 
-- label: Spec Decode Acceptance Rates Nightly
-  key: spec-decode-acceptance-rates-nightly
+- label: Spec Decode AL DFlash Nightly
+  key: spec-decode-dflash-nightly
+  timeout_in_minutes: 90
+  device: h200_35gb
+  optional: true
+  source_file_dependencies:
+    - vllm/v1/spec_decode/
+    - vllm/v1/worker/gpu/spec_decode/
+    - vllm/model_executor/models/qwen3_dflash.py
+    - vllm/model_executor/models/laguna_dflash.py
+    - tests/v1/e2e/spec_decode/
+  commands:
+    - pytest -v -s v1/e2e/spec_decode/acceptance_rates/dflash/
+
+- label: Spec Decode AL DSpark Nightly
+  key: spec-decode-dspark-nightly
   timeout_in_minutes: 60
   device: h200_35gb
   optional: true
   source_file_dependencies:
     - vllm/v1/spec_decode/
     - vllm/v1/worker/gpu/spec_decode/
+    - vllm/model_executor/models/qwen3_dspark.py
+    - vllm/model_executor/models/gemma4_dspark.py
     - tests/v1/e2e/spec_decode/
   commands:
-    - pytest -v -s v1/e2e/spec_decode/test_spec_decode.py -k "acceptance_rates"
+    - pytest -v -s v1/e2e/spec_decode/acceptance_rates/dspark/
+
+- label: Spec Decode AL MTP + Other Acceptance Nightly
+  key: spec-decode-mtp-other-acceptance-nightly
+  timeout_in_minutes: 60
+  device: h200_35gb
+  optional: true
+  source_file_dependencies:
+    - vllm/v1/spec_decode/
+    - vllm/v1/worker/gpu/spec_decode/
+    - tests/v1/e2e/spec_decode/conftest.py
+    - tests/v1/e2e/spec_decode/
+  commands:
+    - pytest -v -s v1/e2e/spec_decode/acceptance_rates/mtp_other/
