+++ tests/models/registry.py
@@ -1449,8 +1449,8 @@ _SPECULATIVE_DECODING_EXAMPLE_MODELS = {
         max_num_seqs=32,
     ),
     "DFlashLagunaForCausalLM": _HfExamplesInfo(
-        "poolside/Laguna-XS-2.1",
-        speculative_model="poolside/Laguna-XS-2.1-DFlash",
+        "poolside/Laguna-XS-2.1-NVFP4",
+        speculative_model="poolside/Laguna-XS-2.1-DFlash-NVFP4",
         use_original_num_layers=True,
         max_model_len=8192,  # Reduce max len to ensure test runs in low-VRAM CI env
         max_num_seqs=32,
