+++ tests/v1/e2e/general/test_kv_sharing_fast_prefill.py
@@ -18,7 +18,7 @@ SEED = 42
 @pytest.fixture
 def test_prompts():
     """
-    Adapted from tests/v1/e2e/spec_decode/test_spec_decode.py
+    Adapted from tests/v1/e2e/spec_decode/utils.py
     """
     prompt_types = ["repeat", "sentence"]
     # Setting higher num prompts increases the chance of numerics mismatch
