+++ tests/v1/e2e/spec_decode/acceptance_rates/mtp_other/__init__.py
@@ -1,0 +1,2 @@
+# SPDX-License-Identifier: Apache-2.0
+# SPDX-FileCopyrightText: Copyright contributors to the vLLM project
